Remove Conditional_Feature interface and rename Interfaces to Contracts
- Remove Conditional_Feature interface (premature feature)
- Remove conditional requirements check from Feature_Loader
- Remove conditional feature implementation from Example_Feature
- Remove conditional feature tests
- Rename Interfaces namespace to Contracts (more commonplace)
- Update all import statements
